Move client remote files services to api/remoteFIles
