+++ client/src/components/FilesDialog/FilesDialog.vue
@@ -5,13 +5,6 @@ import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { BAlert, BButton } from "bootstrap-vue";
 import Vue, { computed, onMounted, ref } from "vue";
 
-import { UrlTracker } from "@/components/DataDialog/utilities";
-import { isSubPath } from "@/components/FilesDialog/utilities";
-import { selectionStates } from "@/components/SelectionDialog/selectionStates";
-import { useConfig } from "@/composables/config";
-import { errorMessageAsString } from "@/utils/simple-error";
-
-import { DirectoryRecord, Model, RecordItem } from "./model";
 import {
     BrowsableFilesSourcePlugin,
     browseRemoteFiles,
@@ -19,7 +12,14 @@ import {
     FilterFileSourcesOptions,
     getFileSources,
     RemoteEntry,
-} from "./services";
+} from "@/api/remoteFiles";
+import { UrlTracker } from "@/components/DataDialog/utilities";
+import { isSubPath } from "@/components/FilesDialog/utilities";
+import { selectionStates } from "@/components/SelectionDialog/selectionStates";
+import { useConfig } from "@/composables/config";
+import { errorMessageAsString } from "@/utils/simple-error";
+
+import { DirectoryRecord, Model, RecordItem } from "./model";
 
 import DataDialogSearch from "@/components/SelectionDialog/DataDialogSearch.vue";
 import DataDialogTable from "@/components/SelectionDialog/DataDialogTable.vue";
