+++ client/src/composables/fileSources.ts
@@ -1,10 +1,6 @@
 import { onMounted, readonly, ref } from "vue";
 
-import {
-    BrowsableFilesSourcePlugin,
-    FilterFileSourcesOptions,
-    getFileSources,
-} from "@/components/FilesDialog/services";
+import { BrowsableFilesSourcePlugin, FilterFileSourcesOptions, getFileSources } from "@/api/remoteFiles";
 
 /**
  * Composable for accessing and working with file sources.
